+++ app/Events/Listeners/InvoiceCreatingListener.php
@@ -31,7 +31,7 @@ class InvoiceCreatingListener
 
         if (!$invoice->due_at)
         {
-            $invoice->due_at = DateFormatter::incrementDateByDays($invoice->invoice_date->format('Y-m-d'), config('fi.invoicesDueAfter'));
+            $invoice->due_at = DateFormatter::incrementDateByDays($invoice->invoice_date->format('Y-m-d'), $invoice->client->client_terms);
         }
 
         if (!$invoice->company_profile_id)