+++ app/Events/Listeners/InvoiceEmailingListener.php
@@ -12,7 +12,7 @@ class InvoiceEmailingListener
         if (config('fi.resetInvoiceDateEmailDraft') and $event->invoice->status_text == 'draft')
         {
             $event->invoice->invoice_date = date('Y-m-d');
-            $event->invoice->due_at       = DateFormatter::incrementDateByDays(date('Y-m-d'), config('fi.invoicesDueAfter'));
+            $event->invoice->due_at       = DateFormatter::incrementDateByDays(date('Y-m-d'),  $event->invoice->client->client_terms);
             $event->invoice->save();
         }
     }