+++ app/Modules/PaymentTerms/Models/Paymentterm.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace FI\Modules\PaymentTerms\Models;
+
+
+use Eloquent;
+
+/**
+ * Class PaymentTerm
+ */
+class PaymentTerm extends Eloquent
+{
+
+    /**
+     * @var bool
+     */
+    public $timestamps = false;
+
+}